requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
requests>=2.31.0

# Optional (Google Sheets datasource)
gspread>=6.0.0
google-auth>=2.0.0